start shell scripts for testing real world usage
